Move `DfsWithPath` to utils
